requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argcomplete==3.5.3
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
avd==0.0.2
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.2.0
boto==2.49.0
Brotli==1.1.0
build==1.2.2.post1
buildozer==1.5.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cmake==3.31.4
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
datasets==3.2.0
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dm-haiku==0.0.13
docutils==0.21.2
EbookLib==0.18
emoji==2.14.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fasteners==0.19
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flask-talisman==1.1.0
flatbuffers==24.12.23
fonttools==4.55.3
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gcs-oauth2-boto-plugin==3.2
google-apitools==0.5.32
google-auth==2.17.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
google-reauth==0.1.1
greenlet==3.1.1
grpcio==1.69.0
gsutil==5.33
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
hf_transfer==0.1.9
hsluv==5.0.4
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflate64==1.0.1
iniconfig==2.0.0
interchange==2021.0.4
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.5
jiter==0.8.2
jmp==0.0.4
joblib==1.4.2
keras==3.8.0
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.4.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
murmurhash==1.0.11
namex==0.0.8
neo4j==5.27.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
odfpy==1.4.1
openai==1.59.7
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pansi==2024.11.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
py2neo==2021.2.4
py7zr==0.22.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybcj==1.0.3
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
PyOpenGL==3.1.7
pyOpenSSL==25.0.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypiwin32==223
pyppmd==1.1.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-for-android==2024.1.21
pytz==2024.2
pyu2f==0.1.5
pywin32==308
PyYAML==6.0.2
pyzstd==0.16.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retry-decorator==1.1.1
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
sh==2.2.1
shell==1.0.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.13.0
SQLAlchemy==2.0.37
srsly==2.5.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
texttable==1.7.0
tf_keras==2.18.0
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.1
vispy==0.14.3
waitress==3.0.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.1
xlrd==2.0.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
